List of changes:
- Divide use-user.ts contents onto multiple files
- Create Session class
