--- context/user/user.ts
@@ -1,4 +1,0 @@
-class User {
-    readonly token: string;
-    setToken: (token: string) => void;
-};
